requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-bootstrap4==2.3.1
django-filter==2.4.0
django-model-utils==4.1.1
django-rest-swagger==2.2.0
djangorestframework==3.12.4
idna==2.10
importlib-metadata==2.1.1
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==8.2.0
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
tqdm==4.60.0
uritemplate==3.0.1
urllib3==1.26.4
zipp==3.4.1
